Added image for the guide
In this guide we explore what our brains are doing when we see an
image. Hence the MRI scan of the human brain with eyes.

I tested this image with the MIT Tool for memorability and it had a
high memorability score of 0.710.

Showed what the tool does in the image.